Add files via upload
The attached file is an image of a Figma project of the design of our website. On the left there is a desktop version of the website with main page and 3 subpages: about us, about the tool, and the tool itself. On the right there is a mobile version of the same website. We might decide that is too much work, so another option is to make just one page with references to paragraphs, where for example clicking on the "about us" tab would transfer us to the "about us" paragraph, within one long page. Give me your thoughts and feedback on that idea.